working on quantization
