+++ Quantization/Cargo.toml
@@ -3,6 +3,10 @@ name = "quant"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+name = "quant"
+path = "src/lib/lib.rs"
+
 [dependencies]
 algo = {path = "../Algorithms"}
 serde = { version = "1.0", features = ["derive"] }
