+++ Simulation/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 algo = {path = "../Algorithms"}
+quant = {path = "../Quantization"}
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 chrono = "0.4"
